+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Changelog
 * Mark categories dangerous in template
 * Restrict */sys/block/.../cid* and */sys/class/.../cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))
 * Restrict system properties ending with *cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))
-* Permanently cache version, log and usage settings
+* Permanently cache XPrivacy version and Android usage settings
 
 **Version 1.10.5 TEST**