+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 * Restrict */sys/block/.../cid* and */sys/class/.../cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))
 * Restrict system properties ending with *cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))
 * Permanently cache XPrivacy version and Android usage settings
+* Do not display number of packages in progress dialog
 
 **Version 1.10.5 TEST**