+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 * Show location for license file in about dialog
 * Fix up navigation (again) ([issue](https://github.com/M66B/XPrivacy/issues/681))
 * Auto import pro license file ([issue](https://github.com/M66B/XPrivacy/issues/703))
+* Allow application specific disable of globally set randomization ([issue](https://github.com/M66B/XPrivacy/issues/706))
 * Restrict */sys/block/.../cid* and */sys/class/.../cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))
 * Restrict system properties ending with *cid* ([issue](https://github.com/M66B/XPrivacy/issues/734))